+++ Java-course-2020/src/app/Exercise_79.java
@@ -1,0 +1,16 @@
+package app;
+
+import java.util.Random;
+public class Exercise_79{
+
+    public static void main (String[]args){
+        Random ra = new Random();
+        double num = ra.nextDouble();
+        System.out.println(num);
+        if (num<0.5){
+            System.out.println("Debajo de la mesa!");
+        } else {
+            System.out.println("Hello friend");
+        }
+    }
+}
